search images rotation

+++ client/src/components/Search/Search.js
@@ -6,6 +6,7 @@ import {Segment, Grid, Divider, Image, Responsive, Button, Container, Checkbox, 
 import tripod from '../../asset/images/tripod-logo.png';
 import API from '../../utils/API';
 import { CloudinaryContext, Transformation } from 'cloudinary-react';
+import isNumeric from 'validator/lib/isNumeric';
 
 class Search extends Component {
 
@@ -29,24 +30,41 @@ class Search extends Component {
       .catch(err => console.log(err));
   };
 
+  // Delete an image
   deleteImage(id, pId) {
-    // // Delete from Cloudinary
-    // API.deleteCloudinary()
-    //   .then(res => console.log("Deleted from Cloudinary"))
-    //   .catch(err => console.log(err));
-    // // Delete form Mongo
-    API.deletePicDetails(id)
-      .then(res => searchImages())//console.log("Image Deleted"))
-      .catch(err => console.log(err));  
+    confirm(" Delete this image ? ");
+    if (confirm("Press a button!") == true) {
+      // // Delete from Cloudinary
+      // API.deleteCloudinary()
+      //   .then(res => console.log("Deleted from Cloudinary"))
+      //   .catch(err => console.log(err));
+      // // Delete form Mongo
+      API.deletePicDetails(id)
+        .then(res => searchImages())//console.log("Image Deleted"))
+        .catch(err => console.log(err));  
+    };
   };
 
-  enlargeImage(key,url){
-    <Popup trigger={<Image key={this.key} />}  style={style}>
-      Hello. This is a regular pop-up which does not allow for lots
-      of content. You cannot fit a lot of words here as the
-      paragraphs will be pretty narrow.
-      <Image src={this.url} />
-    </Popup>
+  // After rotating the image save it in Mongo
+  saveImageChanges(id,newUrl){
+    API.editPicDetails(id , newUrl) 
+      .then(res => searchImages())
+      .catch(err => console.log(err));
+    };
+
+  // Rotate the image
+  rotateImage(id, url, pId) {
+
+    const angle = prompt("Enter your rotation degrees ex: 90  45 -45 -90 180 360");
+    if(isNaN(angle)) {
+      alert("enter a valid number");
+    }
+    else{
+      const change = "a_" + angle;
+      const newUrl = url.substr(0,47) + change + "/" + pId;
+      this.saveImageChanges(id,newUrl);
+    }
+
   };
 
   render() {
@@ -64,13 +82,15 @@ class Search extends Component {
                 <div className='searchText'><h1>FIND YOUR PHOTO'S COLLECTION</h1></div>
                   <CloudinaryContext cloudName="tripod"> 
                     {this.state.gallery.map(data => (
-                      <div className="responsive" key={data._id} style={style.responsive} onClick={ () => this.enlargeImage(data._id,data.pic_url)} >
+                      <div className="responsive" key={data._id} style={style.responsive} >
                         <div className="img" style={style.img} >
                           <Image key={data.pic_url} style={style.displayImage}>
-                            <Popup hoverable flowing size='tiny' position='top center' style={style.popStyle}
-                                trigger={
-                              <Image src = {data.pic_url} style = {style.imgLarge} /> }
-                            >
+                            <CloudinaryContext cloudName="tripod">                    
+                              <Transformation width="200" crop="scale" angle="10" />
+                              <Image publicId={data.pic_public_id}>
+                              <Popup hoverable flowing size='tiny' position='top center' style={style.popStyle}
+                                trigger={<Image src = {data.pic_url} style = {style.imgLarge} />}
+                              >
                               <Container>
                                 <Grid centered columns='one'>
                                   <Grid.Row>
@@ -80,13 +100,16 @@ class Search extends Component {
                                       <Button.Group className='click'>
                                         <Button href={data.pic_url} download={data.pic_url} centered icon='download' basic size='tiny' color='green' />
                                         <Button centered icon='trash' basic size='tiny' color='red' onClick={() => this.deleteImage(data._id , data.pic_public_id)} />
+                                        <Button centered icon='Refresh's basic size='undo' color='blue' onClick={() => this.rotateImage(data._id, data.pic_url, data.pic_public_id)} />
                                       </Button.Group>
                                     </Grid.Column>
                                   </Grid.Row>
                                 </Grid>
                               </Container>
                             </Popup>
-
+                              </Image>
+                            </CloudinaryContext>
+                            
                           </Image>
                         </div>
                       </div>
@@ -111,39 +134,30 @@ class Search extends Component {
 };
 const style = {
   displayImage: {
-    height: 'auto',
-    width: '20%',
+    width: '100%',
+    height: 'auto',    
     padding: '1%',
-    border: '2px solid white',
+    border: '3px solid white',
+    boxShadow:'2px, 5px, 50px black',
     display: 'inline-grid',
     marginTop: '1%',
     marginRight: '1%',
     marginLeft: '1%',
     marginBottom: '1%'
   },
-  // displayImage: {
-  //   width: '20%',
-  //   height: 'auto',
-  //   padding:'2%',
-  //   border:'2px solid white',
-  //   display:'inline-grid',
-  //   marginTop:'1%',
-  //   marginRight:'1%',
-  //   marginLeft:'1%',
-  //   marginBottom:'1%',
-  // },
+  popStyle: {
+  backgroundColor:'rgba(0,0,0,0.0)',
+  border:'none',
+  boxShadow:'none'
+  },
   img:{
     border: '1px solid #ccc'
   },
   responsive: {
-    padding:'0 6px',
+    padding:'0px',
     float:'left',
-    width:'24%',
+    width:'20%',
     margin:'10px'
-  },
-  desc: {
-    padding: '15px',
-    textAlign:'center'
   }
 };
 export default Search;
